1.Global Parser Instantiation (for the purposes of calling parser.arg_parse() from different python modules; and thus enabling us to return the dataframe); 2.Validation changed, since it is redundant with the introduction of the argparser. 3. Added functionality of returning a dataframe from another python module through the following: import coinmarketcap_usd_history; df = main(['bitcoin','2017','2017','--toDf']);.
